requirements.txt
yfinance>=0.2.54
pandas>=2.2.3
numpy>=2.2.4
PyYAML>=6.0.2
Flask>=3.1.0
gunicorn>=23.0.0
Pillow>=11.1.0
rapidocr-onnxruntime>=1.3.25
